Add files via upload
The dataset contains a single CSV file, movies.csv, with the following columns:

id: Unique identifier for each movie.
title: Title of the movie.
director: Name of the movie director.
cast: Names of the movie's main cast members.
country: Country of origin of the movie.
release_year: Year the movie was released.
genre: Genre of the movie.
duration_min: Duration of the movie in minutes.
rating: Average rating of the movie on a scale of 0 to 10.
votes: Number of votes for the movie on the rating platform.

This dataset can be used for a variety of purposes, such as analyzing trends in movie genres and ratings, studying the work of particular directors, or building recommendation systems. The dataset is provided under the Creative Commons Attribution-ShareAlike 3.0 Unported License, which allows for free use and distribution of the data as long as the original source is attributed and any derivative works are shared under the same license.